Add new helper methods to use PsiAstTemplate
Summary:
I want to make templates more prominent in codemodding.
One issue is that I find the current API I build to not be as convenient, it requires knowing a bunch of Psi classes to make it start working, which I know, but is bad for ramp up.

These straightforward methods should make things nicer.

Reviewed By: jsendros

Differential Revision: D53857236

fbshipit-source-id: f3d757f12f5db21d6d5df0098ac614f25b86eaeb