Replace strategy to match lists of arguments
Summary:
The previous startegy to match lists was very opinionated about indices, this is not good when we move to optional matchers. So let's replace it, and simplify things.

Adding a new list matching method which works for optional matchers after non optional ones, and verifies the size makes sense. Matching any number of variables is now no longer supported, but we can bring it back if/when we need to and decide how to expose it.

Reviewed By: jsendros

Differential Revision: D54130332

fbshipit-source-id: 71335c77d48848023d7fb5e33928192c8f9a4339