Add variable for text in templates
Summary:
We want to allow templates to support a few basic predicates on their variables.
The most import that comes to mind is checking the type of the expression, but we will start with a simpler option to match on a regex for the text which is also very useful.

The syntax is in the variable as in `#a{text=MyRegex}#`, this is inspired by how IntelliJ's structural replace presents similar information and I prefer to keep it consistent with other offerings.

The code coould use some refactoring and cleanups in the way we do variables.

Reviewed By: jsendros

Differential Revision: D54292632

fbshipit-source-id: 322ddef77ec215999ae244ba5b8dce3105617d90