Fix issues when matching on method calls with variables in Java
Summary:
This came out pretty complicated.
A `PsiReferenceExpression` can split into two `PsiReferenceExpression` or an `PsiReferenceExpression` and `PsiIdentifier`. And we need to deal a bunch of different cases, since if the user gives us `a.#b#` we need to load a variable that needs to match either a `PsiIdentifier` or `PsiReferenceExpression` and this means we need to a trick we we build a matcher that looks for `PsiElement` and then verifies the type.

I don't know of an easier way to do it for now.

Reviewed By: jsendros

Differential Revision: D55351567

fbshipit-source-id: 2066865ed2d70b26444b25d50ac1898509096d13