Rank and resort matcher functions to avoid doing expensive checks first
Summary:
Once we add type resolving in some places we might have super slow actions in the matcher. In those cases, we want faster actions to run fast.

The ideal formula would be based on the average run time  and average success rate (at stoping the matcher), I was too lazy to do that, so I just sort by time, and average it with all previous runs. This does the job, as the test shows.

Reviewed By: jsendros

Differential Revision: D58187861

fbshipit-source-id: 7faa5cbe62720d4b94dab7feefd477a564048fc9