Preliminary prototype of a utility to find usages of variables
Summary:
This has been a common theme, it's not hard to do, but it'a bunch of code - and the edge cases... ok, it is hard to do. One has to handle `this`, labels, function references, order of initialization and more.

`DeclarationFinder` actually does a bunch of it, so we could just use that, and transform the API to a more useful one.

I'll add more to it to work in Java, and we have more general elements.

Differential Revision: D63765290

fbshipit-source-id: 66c0b0a4b58c150bf832ca92eb8d4848a3cc8fc3