Improve UsagesFinder to differentiate functions and variables better
Summary:
This improves the code with a bunch of needed complexity - we can better handle functions and variables that collide on their name, which is of course ok and legit in Java and Kotlin.

We still don't handle this and super pointers really well, but using this in codemods will be better than the ad hoc stuff people do. (and shorter)

Reviewed By: jsendros

Differential Revision: D64639579

fbshipit-source-id: 497ea7ef51e1135bab818fd775a02f17d8090302