consider PsiTypeParameters before PsiClass
Summary: `T extends Foo` is a `PsiTypeParameter` which are a subclass of `PsiClass`. This lead to an issue when calling `collectDescendantsOfType<AClassOrObject>()` where it would return `T extends Foo` as `AClassOrObject`.  That is obviously not a class so we add the check on PsiTypeParameter before PsiClass to ensure we do not create `AClassOrObject`

Reviewed By: lcharriere

Differential Revision: D80682461

fbshipit-source-id: c6db92b096a19f2b2995a26e37b5159d65a9978a