Intro AFile.replaceInPlace
Summary:
In an effort to break up codemods into parts that could be managed DevMate I want to extract common actions into reusable methods.
This is the first method called "replace" that replaces all elements of type T with a given replacement. It also adds neccessary deps and imports.

As an example I migrated MigrateSpykPlugin to use this new method

Differential Revision: D81163672

fbshipit-source-id: e67de700482951870045ed855713b02831bfe54a